Fix GetRequiredSpace calculations
GetRequiredSpace uses wrong math to calculate overhead expansion, which usually causes inflated space to be significantly larger than it should.

This commit fixes this calculation and removes the redundant CalculateOverheadSpace.
